requirements.txt
Django==4.2.11
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
